Task2

+++ Tyuiu.IvanovPG.Sprint1.Task2.V9.Lib/DataService.cs
@@ -4,7 +4,7 @@ namespace Tyuiu.IvanovPG.Sprint1.Task2.V0.Lib
 {
     public class DataService : ISprint1Task2V9
     {
-        public double CalculateVolumeCircle(double r)
+        public double CalculateVolumeCircle(int r)
         {
             return (4/3) * 3.14 * r * r * r;
         }
